Share code between sql and Dask queries
