+++ gpu_bdb/bdb_tools/q27_utils.py
@@ -1,0 +1,31 @@
+#
+# Copyright (c) 2019-2022, NVIDIA CORPORATION.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import spacy
+
+def ner_parser(df, col_string, batch_size=256):
+    spacy.require_gpu()
+    nlp = spacy.load("en_core_web_sm")
+    docs = nlp.pipe(df[col_string], disable=["tagger", "parser"], batch_size=batch_size)
+    out = []
+    for doc in docs:
+        l = [ent.text for ent in doc.ents if ent.label_ == "ORG"]
+        val = ", "
+        l = val.join(l)
+        out.append(l)
+    df["company_name_list"] = out
+    return df
+
