+++ gpu_bdb/queries/q05/gpu_bdb_query_05.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2019-2020, NVIDIA CORPORATION.
+# Copyright (c) 2019-2022, NVIDIA CORPORATION.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,7 +14,6 @@
 # limitations under the License.
 #
 
-import sys
 import os
 import glob
 
@@ -25,14 +24,13 @@ from bdb_tools.utils import (
 )
 
 from bdb_tools.readers import build_reader
-from bdb_tools.cupy_metrics import cupy_precision_score
+from bdb_tools.q05_utils import build_and_predict_model
 
 import cupy as cp
 import numpy as np
 from dask import delayed
 import dask
 import pandas as pd
-from sklearn.metrics import roc_auc_score
 
 #
 # Query Configuration
@@ -44,14 +42,6 @@ wcs_columns = ["wcs_item_sk", "wcs_user_sk"]
 items_columns = ["i_item_sk", "i_category", "i_category_id"]
 customer_columns = ["c_customer_sk", "c_current_cdemo_sk"]
 customer_dem_columns = ["cd_demo_sk", "cd_gender", "cd_education_status"]
-
-# Logistic Regression params
-# solver = "LBFGS" Used by passing `penalty=None` or "l2"
-# step_size = 1 Not used
-# numCorrections = 10 Not used
-iterations = 100
-C = 10_000  # reg_lambda = 0 hence C for model is a large value
-convergence_tol = 1e-9
 
 
 def read_tables(config):
@@ -70,47 +60,6 @@ def read_tables(config):
     )
 
     return (item_ddf, customer_ddf, customer_dem_ddf)
-
-
-def build_and_predict_model(ml_input_df):
-    """
-    Create a standardized feature matrix X and target array y.
-    Returns the model and accuracy statistics
-    """
-    import cuml
-    from cuml.metrics import confusion_matrix
-
-    feature_names = ["college_education", "male"] + [
-        "clicks_in_%d" % i for i in range(1, 8)
-    ]
-    X = ml_input_df[feature_names]
-    # Standardize input matrix
-    X = (X - X.mean()) / X.std()
-    y = ml_input_df["clicks_in_category"]
-
-    model = cuml.LogisticRegression(
-        tol=convergence_tol,
-        penalty="none",
-        solver="qn",
-        fit_intercept=True,
-        max_iter=iterations,
-        C=C,
-    )
-    model.fit(X, y)
-    #
-    # Predict and evaluate accuracy
-    # (Should be 1.0) at SF-1
-    #
-    results_dict = {}
-    y_pred = model.predict(X)
-
-    results_dict["auc"] = roc_auc_score(y.to_array(), y_pred.to_array())
-    results_dict["precision"] = cupy_precision_score(cp.asarray(y), cp.asarray(y_pred))
-    results_dict["confusion_matrix"] = confusion_matrix(
-        cp.asarray(y, dtype="int32"), cp.asarray(y_pred, dtype="int32")
-    )
-    results_dict["output_type"] = "supervised"
-    return results_dict
 
 
 def get_groupby_results(file_list, item_df):
