+++ gpu_bdb/queries/q05/gpu_bdb_query_05_dask_sql.py
@@ -14,13 +14,9 @@
 # limitations under the License.
 #
 
-import sys
-
 from bdb_tools.cluster_startup import attach_to_cluster
-from dask_cuda import LocalCUDACluster
-from dask.distributed import Client, wait
+from dask.distributed import wait
 from dask import delayed
-import os
 
 from bdb_tools.utils import (
     benchmark,
@@ -28,18 +24,8 @@ from bdb_tools.utils import (
     run_query,
 )
 from bdb_tools.readers import build_reader
-from bdb_tools.cupy_metrics import cupy_precision_score
-from sklearn.metrics import roc_auc_score
-import cupy as cp
+from bdb_tools.q05_utils import build_and_predict_model
 
-
-# Logistic Regression params
-# solver = "LBFGS" Used by passing `penalty=None` or "l2"
-# step_size = 1 Not used
-# numCorrections = 10 Not used
-iterations = 100
-C = 10_000  # reg_lambda = 0 hence C for model is a large value
-convergence_tol = 1e-9
 
 wcs_columns = ["wcs_item_sk", "wcs_user_sk"]
 items_columns = ["i_item_sk", "i_category", "i_category_id"]
@@ -68,47 +54,6 @@ def read_tables(data_dir, c, config):
     c.create_table("customer", customer_ddf, persist=False)
     c.create_table("item", item_ddf, persist=False)
     c.create_table("customer_demographics", customer_dem_ddf, persist=False)
-
-
-def build_and_predict_model(ml_input_df):
-    """
-    Create a standardized feature matrix X and target array y.
-    Returns the model and accuracy statistics
-    """
-    import cuml
-    from cuml.metrics import confusion_matrix
-
-    feature_names = ["college_education", "male"] + [
-        "clicks_in_%d" % i for i in range(1, 8)
-    ]
-    X = ml_input_df[feature_names]
-    # Standardize input matrix
-    X = (X - X.mean()) / X.std()
-    y = ml_input_df["clicks_in_category"]
-
-    model = cuml.LogisticRegression(
-        tol=convergence_tol,
-        penalty="none",
-        solver="qn",
-        fit_intercept=True,
-        max_iter=iterations,
-        C=C,
-    )
-    model.fit(X, y)
-    #
-    # Predict and evaluate accuracy
-    # (Should be 1.0) at SF-1
-    #
-    results_dict = {}
-    y_pred = model.predict(X)
-
-    results_dict["auc"] = roc_auc_score(y.to_array(), y_pred.to_array())
-    results_dict["precision"] = cupy_precision_score(cp.asarray(y), cp.asarray(y_pred))
-    results_dict["confusion_matrix"] = confusion_matrix(
-        cp.asarray(y, dtype="int32"), cp.asarray(y_pred, dtype="int32")
-    )
-    results_dict["output_type"] = "supervised"
-    return results_dict
 
 
 def main(data_dir, client, c, config):
