+++ gpu_bdb/queries/q20/gpu_bdb_query_20.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2019-2020, NVIDIA CORPORATION.
+# Copyright (c) 2019-2022, NVIDIA CORPORATION.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,11 +14,7 @@
 # limitations under the License.
 #
 
-import sys
-import cupy as cp
-import rmm
 import numpy as np
-
 
 from bdb_tools.utils import (
     benchmark,
@@ -27,14 +23,9 @@ from bdb_tools.utils import (
     run_query,
 )
 from bdb_tools.readers import build_reader
+from bdb_tools.q20_utils import get_clusters
 from dask import delayed
 from dask.distributed import wait
-
-
-# q20 parameters
-N_CLUSTERS = 8
-CLUSTER_ITERATIONS = 20
-N_ITER = 5
 
 
 def read_tables(config):
@@ -62,31 +53,6 @@ def read_tables(config):
         "store_returns", relevant_cols=store_returns_cols
     )
     return store_sales_df, store_returns_df
-
-
-def get_clusters(client, ml_input_df, feature_cols):
-    """
-    Takes the dask client, kmeans_input_df and feature columns.
-    Returns a dictionary matching the output required for q20
-    """
-    import dask_cudf
-
-    ml_tasks = [
-        delayed(train_clustering_model)(df, N_CLUSTERS, CLUSTER_ITERATIONS, N_ITER)
-        for df in ml_input_df[feature_cols].to_delayed()
-    ]
-
-    results_dict = client.compute(*ml_tasks, sync=True)
-
-    labels = results_dict["cid_labels"]
-
-    labels_final = dask_cudf.from_cudf(labels, npartitions=ml_input_df.npartitions)
-    ml_input_df["label"] = labels_final.reset_index()[0]
-
-    output = ml_input_df[["user_sk", "label"]]
-
-    results_dict["cid_labels"] = output
-    return results_dict
 
 
 def remove_inf_and_nulls(df, column_names, value=0.0):
