+++ gpu_bdb/queries/q20/gpu_bdb_query_20_dask_sql.py
@@ -26,41 +26,13 @@ from bdb_tools.utils import (
     benchmark,
     gpubdb_argparser,
     run_query,
-    train_clustering_model
 )
 
 from bdb_tools.readers import build_reader
 
+from bdb_tools.q20_utils import get_clusters
+
 from dask_sql import Context
-
-# q20 parameters
-N_CLUSTERS = 8
-CLUSTER_ITERATIONS = 20
-N_ITER = 5
-
-
-def get_clusters(client, ml_input_df, feature_cols):
-    """
-    Takes the dask client, kmeans_input_df and feature columns.
-    Returns a dictionary matching the output required for q20
-    """
-    import dask_cudf
-    ml_tasks = [
-        delayed(train_clustering_model)(df, N_CLUSTERS, CLUSTER_ITERATIONS, N_ITER)
-        for df in ml_input_df[feature_cols].to_delayed()
-    ]
-
-    results_dict = client.compute(*ml_tasks, sync=True)
-
-    labels = results_dict["cid_labels"]
-
-    labels_final = dask_cudf.from_cudf(labels, npartitions=ml_input_df.npartitions)
-    ml_input_df["label"] = labels_final.reset_index()[0]
-
-    output = ml_input_df[["user_sk", "label"]]
-
-    results_dict["cid_labels"] = output
-    return results_dict
 
 
 def read_tables(data_dir, c, config):
